// src/markup/interfaces/elements.interface.ts
export interface IControlElements {
  controlWrapper: HTMLDivElement | null;
  controlProgress: HTMLDivElement | null;
  controlHandle: HTMLDivElement | null;
  controlVolume: HTMLDivElement | null;
}

export interface IBandcampElements {
  bcContainer: HTMLDivElement | null;
  bcWrapper: HTMLDivElement | null;
}